requirements.txt
azure_storage==0.37.0
click==8.1.7
dlib==19.24.4
face_recognition_models==0.3.0
Flask==3.0.3
mediapipe==0.10.14
numpy==1.26.4
openai==1.44.1
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
picamera==1.13
Pillow==10.4.0
scikit_learn==1.5.1
scipy==1.14.1
setuptools==74.1.2
